+++ ledger-service/http-json-testing/BUILD.bazel
@@ -57,7 +57,6 @@ hj_scalacopts = lf_scalacopts + [
             "//ledger/ledger-runner-common",
             "//ledger/metrics",
             "//ledger/participant-integration-api",
-            "//ledger/sandbox-common",
             "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
             "//ledger/sandbox-on-x",
             "//ledger/sandbox-on-x:sandbox-classic-server",
