+++ ledger-service/http-json/BUILD.bazel
@@ -543,7 +543,6 @@ test_suite(
             "//ledger/metrics",
             "//ledger/participant-integration-api",
             "//ledger/participant-integration-api:participant-integration-api-tests-lib",
-            "//ledger/sandbox-common",
             "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
             "//ledger/sandbox-on-x",
             "//ledger/sandbox-on-x:sandbox-classic-server",
