+++ ledger/sandbox-on-x/BUILD.bazel
@@ -149,10 +149,8 @@ da_scala_library(
         "//ledger/ledger-api-domain",
         "//ledger/ledger-resources",
         "//ledger/ledger-runner-common",
-        "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-integration-api:participant-integration-api-tests-lib",
-        "//ledger/sandbox-common",
         "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
         "//ledger/test-common",
         "//libs-scala/grpc-utils",
@@ -316,10 +314,8 @@ da_scala_library(
         "//ledger/ledger-resources",
         "//ledger/ledger-resources:ledger-resources-test-lib",
         "//ledger/ledger-runner-common",
-        "//ledger/metrics",
         "//ledger/participant-integration-api",
         "//ledger/participant-integration-api:participant-integration-api-tests-lib",
-        "//ledger/sandbox-common",
         "//ledger/sandbox-common:sandbox-common-scala-tests-lib",
         "//ledger/test-common",
         "//libs-scala/contextualized-logging",
@@ -411,7 +407,7 @@ alias(
 )
 
 da_scala_binary(
-    name = "sandbox-on-x-ephemeral-postgresql",
+    name = "sandbox-on-x-ephemeral-oracle",
     main_class = "com.daml.ledger.sandbox.MainWithEphemeralOracleUser",
     visibility = ["//visibility:public"],
     runtime_deps = [],
@@ -421,7 +417,7 @@ da_scala_binary(
 )
 
 da_scala_binary(
-    name = "sandbox-on-x-ephemeral-oracle",
+    name = "sandbox-on-x-ephemeral-postgresql",
     main_class = "com.daml.ledger.sandbox.MainWithEphemeralPostgresql",
     visibility = ["//visibility:public"],
     runtime_deps = [],
